backend/search/requirements.txt
fastapi
uvicorn
aiohttp
pydantic-settings
azure-search-documents
openai
pandas